feat: Centralize device value formatting and increment version
This commit refactors the application to use a unified utility for device value formatting and updates the project version.

- **Value Formatting:**
  - Introduced `formatDeviceValue` in `DeviceValueFormatter.kt` to centralize rendering logic for card and detail screens.
  - Replaced local, redundant implementations in `DeviceDetailScreen.kt` and `GreenhouseDetailScreen.kt` with the new shared utility.
  - Logic now prioritizes the `dataType` metadata to determine if a value is boolean, ensuring numeric sensors (like lux or UV index) are not incorrectly rendered as "OFF".
  - Enhanced numeric formatting: values ≥ 10,000 are now abbreviated (e.g., `45K`), and whole-number doubles are displayed as integers.

- **Versioning:**
  - Incremented `versionCode` to `4` and `versionName` to `0.2.0` in `composeApp/build.gradle.kts`.
