+++ composeApp/build.gradle.kts
@@ -149,8 +149,8 @@ android {
         applicationId = "com.apptolast.greenhousefronts"
         minSdk = libs.versions.android.minSdk.get().toInt()
         targetSdk = libs.versions.android.targetSdk.get().toInt()
-        versionCode = 3
-        versionName = "0.1.2"
+        versionCode = 4
+        versionName = "0.2.0"
     }
     buildFeatures {
         buildConfig = true
